Convert LinearAssembler tests to work with the cython-wrapped LinearAssembler
