+++ tests/test_assembly.py
@@ -51,473 +51,11 @@ from . import khmer_tst_utils as utils
 import pytest
 import screed
 
+from .graph_features import *
+from .graph_features import K
 
 def teardown():
     utils.cleanup()
-
-# We just define this globally rather than in a module-level fixture,
-# as we need it during parameterization and whatnot.
-K = 21
-
-
-class Kmer(str):
-
-    def __init__(self, value, pos=0):
-        self.pos = pos
-
-    def __new__(cls, value, pos=0):
-        if not len(value) == K:
-            raise ValueError('bad k-mer length')
-        return str.__new__(cls, value)
-
-
-def mutate_base(base):
-    if base in 'AT':
-        return random.choice('GC')
-    elif base in 'GC':
-        return random.choice('AT')
-    else:
-        assert False, 'bad base'
-
-
-def mutate_sequence(sequence, N=1):
-    sequence = list(sequence)
-    positions = random.sample(range(len(sequence)), N)
-
-    for i in positions:
-        sequence[i] = mutate_base(sequence[i])
-
-    return ''.join(sequence)
-
-
-def mutate_position(sequence, pos):
-    sequence = list(sequence)
-    sequence[pos] = mutate_base(sequence[pos])
-    return ''.join(sequence)
-
-
-def get_random_sequence(length, exclude=None):
-    '''Generate a random (non-looping) nucleotide sequence.
-
-    To be non-overlapping, the sequence should not include any repeated
-    length K-1 k-mers.
-
-    Args:
-        exclude (str): If not None, add the k-mers from this sequence to the
-        seen set.
-
-    Returns:
-        str: A random non-looping sequence.
-    '''
-
-    seen = set()
-
-    def add_seen(kmer):
-        seen.add(kmer)
-        seen.add(revcomp(kmer))
-
-    if exclude is not None:
-        for pos in range(0, len(exclude) - K):
-            add_seen(exclude[pos:pos + K - 1])
-
-    seq = [random.choice('ACGT') for _ in range(K - 1)]  # do first K-1 bases
-    add_seen(''.join(seq))
-
-    while(len(seq) < length):
-        next_base = random.choice('ACGT')
-        next_kmer = ''.join(seq[-K + 2:] + [next_base])
-        assert len(next_kmer) == K - 1
-        if (next_kmer) not in seen:
-            seq.append(next_base)
-            add_seen(next_kmer)
-        else:
-            continue
-    return ''.join(seq)
-
-
-def reads(sequence, L=100, N=100):
-    positions = list(range(len(sequence) - L))
-    for i in range(N):
-        start = random.choice(positions)
-        yield sequence[start:start + L]
-
-
-def kmers(sequence):
-    for i in range(len(sequence) - K + 1):
-        yield sequence[i:i + K]
-
-
-def test_mutate_sequence():
-    for _ in range(100):
-        assert 'A' not in mutate_sequence('A' * 10, 10)
-        assert 'T' not in mutate_sequence('T' * 10, 10)
-        assert 'C' not in mutate_sequence('C' * 10, 10)
-        assert 'G' not in mutate_sequence('G' * 10, 10)
-
-
-def test_mutate_position():
-    assert mutate_position('AAAA', 2) in ['AACA', 'AAGA']
-    assert mutate_position('TTTT', 2) in ['TTCT', 'TTGT']
-    assert mutate_position('CCCC', 2) in ['CCAC', 'CCTC']
-    assert mutate_position('GGGG', 2) in ['GGAG', 'GGTG']
-
-
-def test_reads():
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-
-    for read in reads(contig):
-        assert read in contig
-
-    for read in reads(contig):
-        assert mutate_sequence(read) not in contig
-
-
-'''
-# GRAPH STRUCTURE FIXTURES
-
-These fixtures emit various graph structures with their corresponding
-sequences and important nodes. They take a random sequence fixture and
-a graph fixture, then consume sequence and generate k-mers accordingly.
-
-We're using a bespoke but simple language to describe graph structures in the
-docstrings of these tests. It is as follows:
-
-    o: Node
-    [x:y]: Node at position in sequence
-    [x:y]+S: Node at position in sequence with extra base (where S in ACGT)
-    (Name), ([x:y] Name): Named node, named node at position
-    → : Edge
-    ~~: Tandem →o→ repeats
-'''
-
-
-@pytest.fixture(params=['simple-genome.fa'])
-def known_sequence(request):
-    fn = utils.get_test_data(request.param)
-    return list(screed.open(fn))[0].sequence
-
-
-@pytest.fixture(params=list(range(500, 1600, 500)),
-                ids=lambda val: '(L={0})'.format(val))
-def random_sequence(request):
-
-    def get(exclude=None):
-        return get_random_sequence(request.param, exclude=exclude)
-
-    return get
-
-
-@pytest.fixture(params=[khmer.Nodegraph, khmer.Countgraph],
-                ids=['(Type=Nodegraph)', '(Type=Countgraph)'])
-def graph(request):
-
-    num_kmers = 50000
-    des_fp = 0.00001
-    args = optimal_fp(num_kmers, des_fp)
-    print('Graph Params:', args)
-
-    return request.param(K, args.htable_size, args.num_htables)
-
-
-def hdn_counts(sequence, graph):
-    '''Get the degree distribution of nodes with degree more than 2.
-    '''
-
-    hdns = {}
-    for kmer in kmers(sequence):
-        d = graph.kmer_degree(kmer)
-        if d > 2:
-            hdns[d] = hdns.get(d, 0) + 1
-
-    return hdns
-
-
-@pytest.fixture
-def linear_structure(request, graph, random_sequence):
-    '''Sets up a simple linear path graph structure.
-
-    sequence
-    [0]→o→o~~o→o→[-1]
-    '''
-    sequence = random_sequence()
-    graph.consume(sequence)
-
-    # Check for false positive neighbors in our graph
-    # Mark as an expected failure if any are found
-    if hdn_counts(sequence, graph):
-        request.applymarker(pytest.mark.xfail)
-
-    return graph, sequence
-
-
-@pytest.fixture(params=[K * 2, -K * 2],
-                ids=['(Where={0})'.format(i) for i in ['Start', 'End']])
-def right_tip_structure(request, graph, random_sequence):
-    '''
-    Sets up a graph structure like so:
-                                 ([S+1:S+K]+B tip)
-    sequence                   ↗
-    [0]→o→o~~o→(L)→([S:S+K] HDN)→(R)→o→o→o~~o→[-1]
-
-    Where S is the start position of the high degreen node (HDN).
-    That is, it has a single branch at the Sth K-mer.
-    '''
-    sequence = random_sequence()
-    S = request.param
-    if S < 0:
-        S = len(sequence) + S
-    # the HDN
-    HDN = Kmer(sequence[S:S + K], pos=S)
-    # left of the HDN
-    L = Kmer(sequence[S - 1:S - 1 + K], pos=S - 1)
-    # right of the HDN
-    R = Kmer(sequence[S + 1:S + 1 + K], pos=S + 1)
-    # the branch kmer
-    tip = Kmer(mutate_position(R, -1),
-               pos=R.pos)
-
-    graph.consume(sequence)
-    graph.count(tip)
-
-    # Check for false positive neighbors and mark as expected failure if found
-    if hdn_counts(sequence, graph) != {3: 1}:
-        request.applymarker(pytest.mark.xfail)
-
-    return graph, sequence, L, HDN, R, tip
-
-
-@pytest.fixture(params=[K * 2, -K * 2],
-                ids=['(Where={0})'.format(i) for i in ['Start', 'End']])
-def right_double_fork_structure(request, linear_structure, random_sequence):
-    '''
-    Sets up a graph structure like so:
-                                               branch
-                                 ([S+1:S+K]+B)→o~~o→o
-    core_sequence               ↗
-    [0]→o→o~~o→(L)→([S:S+K] HDN)→(R)→o→o→o~~o→[-1]
-
-    Where S is the start position of the high degreen node (HDN)
-    and B is the mutated base starting the branch.
-    '''
-
-    graph, core_sequence = linear_structure
-    print('\nCore Len:', len(core_sequence))
-    branch_sequence = random_sequence(exclude=core_sequence)
-    print('Branch len:', len(branch_sequence))
-
-    # start position of the HDN
-    S = request.param
-    if S < 0:
-        S = len(core_sequence) + S
-    # the HDN
-    HDN = Kmer(core_sequence[S:S + K], pos=S)
-    # left of the HDN
-    L = Kmer(core_sequence[S - 1:S - 1 + K], pos=S - 1)
-    # right of the HDN
-    R = Kmer(core_sequence[S + 1:S + 1 + K], pos=S + 1)
-    # the branch sequence, mutated at position S+1
-    branch_start = core_sequence[:R.pos] + mutate_position(R, -1)
-    branch_sequence = branch_start + branch_sequence
-
-    graph.consume(core_sequence)
-    graph.consume(branch_sequence)
-
-    # Check for false positive neighbors and mark as expected failure if found
-    core_hdns = hdn_counts(core_sequence, graph)
-    branch_hdns = hdn_counts(branch_sequence, graph)
-
-    # the core and branch sequences should each have exactly
-    # ONE node of degree 3 (HDN)
-    if core_hdns != {3: 1} or branch_hdns != {3: 1}:
-        print(core_hdns, branch_hdns)
-        request.applymarker(pytest.mark.xfail)
-
-    return graph, core_sequence, L, HDN, R, branch_sequence
-
-
-@pytest.fixture
-def right_triple_fork_structure(request, right_double_fork_structure,
-                                random_sequence):
-    '''
-    Sets up a graph structure like so:
-
-                                       top_branch
-                                ([:S+1]+B)→o~~o→o
-    core_sequence              ↗
-    [0]→o→o~~o→(L)→([S:S+K] HDN)→(R)→o→o→o~~o→[-1]
-                               ↘
-                                ([:S+1]+B)→o~~o→o
-                                     bottom_branch
-
-    Where S is the start position of the high degreen node (HDN).
-    '''
-
-    graph, core_sequence, L, HDN, R, top_sequence = right_double_fork_structure
-    bottom_branch = random_sequence(exclude=core_sequence + top_sequence)
-    print(len(core_sequence), len(top_sequence), len(bottom_branch))
-
-    # the branch sequence, mutated at position S+1
-    # choose a base not already represented at that position
-    bases = {'A', 'C', 'G', 'T'}
-    mutated = random.choice(list(bases - {R[-1], top_sequence[R.pos + K - 1]}))
-
-    bottom_sequence = core_sequence[:HDN.pos + K] + mutated + bottom_branch
-
-    graph.consume(bottom_sequence)
-
-    # Check for false positive neighbors and mark as expected failure if found
-    core_hdns = hdn_counts(core_sequence, graph)
-    top_hdns = hdn_counts(top_sequence, graph)
-    bottom_hdns = hdn_counts(bottom_sequence, graph)
-
-    # the core, top, and bottom sequences should each have exactly
-    # ONE node of degree 4 (HDN)
-    if not (core_hdns == top_hdns == bottom_hdns == {4: 1}):
-        print(core_hdns, top_hdns, bottom_hdns)
-        request.applymarker(pytest.mark.xfail)
-
-    return graph, core_sequence, L, HDN, R, top_sequence, bottom_sequence
-
-
-@pytest.fixture(params=[K * 2, -K * 2],
-                ids=['(Where={0})'.format(i) for i in ['Start', 'End']])
-def left_tip_structure(request, graph, random_sequence):
-    '''
-    Sets up a graph structure like so:
-
-    branch
-    (B+[S:S+K-1] tip)
-                     ↘                    sequence
-        [0]→o~~o→(L)→([S:S+K] HDN)→(R)→o→o~~o→[-1]
-
-    Where S is the start position of the HDN.
-    '''
-    sequence = random_sequence()
-    S = request.param
-    if S < 0:
-        S = len(sequence) + S
-    tip = Kmer(mutate_position(sequence[S - 1:S - 1 + K], 0),
-               pos=S - 1 + K)
-    HDN = Kmer(sequence[S:S + K], pos=S)
-    L = Kmer(sequence[S - 1:S - 1 + K], pos=S - 1)
-    R = Kmer(sequence[S + 1:S + 1 + K], pos=S + 1)
-
-    graph.consume(sequence)
-    graph.count(tip)
-
-    # Check for false positive neighbors and mark as expected failure if found
-    if hdn_counts(sequence, graph) != {3: 1}:
-        request.applymarker(pytest.mark.xfail)
-
-    return graph, sequence, L, HDN, R, tip
-
-
-@pytest.fixture(params=[K * 2, -K * 2],
-                ids=['(Where={0})'.format(i) for i in ['Start', 'End']])
-def left_double_fork_structure(request, linear_structure, random_sequence):
-    '''
-    Sets up a graph structure like so:
-
-    o→o~~o→(B+[S:S+K-1])
-                        ↘                  core_sequence
-          [0]→o→o~~o→(L)→([S:S+K] HDN)→(R)→o→o→o~~o→[-1]
-
-    Where S is the start position of the high degreen node (HDN).
-    '''
-
-    graph, core_sequence = linear_structure
-    branch_sequence = random_sequence(exclude=core_sequence)
-
-    # start position of the HDN
-    S = request.param
-    if S < 0:
-        S = len(core_sequence) + S
-    # the HDN
-    HDN = Kmer(core_sequence[S:S + K], pos=S)
-    # left of the HDN
-    L = Kmer(core_sequence[S - 1:S - 1 + K], pos=S - 1)
-    # right of the HDN
-    R = Kmer(core_sequence[S + 1:S + 1 + K], pos=S + 1)
-    # the branch sequence, mutated at position 0 in L,
-    # whih is equivalent to the K-1 prefix of HDN prepended with a new base
-    branch_start = mutate_position(L, 0)
-    branch_sequence = branch_sequence + \
-        branch_start + core_sequence[L.pos + K:]
-
-    graph.consume(core_sequence)
-    graph.consume(branch_sequence)
-
-    # Check for false positive neighbors and mark as expected failure if found
-    core_hdns = hdn_counts(core_sequence, graph)
-    branch_hdns = hdn_counts(branch_sequence, graph)
-
-    # the core and branch sequences should each have exactly
-    # ONE node of degree 3 (HDN)
-    if not (core_hdns == branch_hdns == {3: 1}):
-        request.applymarker(pytest.mark.xfail)
-
-    return graph, core_sequence, L, HDN, R, branch_sequence
-
-
-@pytest.fixture(params=[K * 2, (-K * 2) - 2],
-                ids=['(Where={0})'.format(i) for i in ['Start', 'End']])
-def snp_bubble_structure(request, linear_structure):
-    '''
-    Sets up a graph structure resulting from a SNP (Single Nucleotide
-    Polymorphism).
-
-                        (HDN_L[1:]+SNP)→o~~o→(SNP+)
-                      ↗                            ↘
-    o~~([S:S+K] HDN_L)                             ([S+K+1:S+2K+1] HDN_R)~~o
-                      ↘                           ↗
-                        (HDN_L[1:]+W)→o~~o~~o→(W+)
-
-    Where S is the start position of HDN directly left of the SNP (HDN_L),
-    SNP is the mutated base, and W is the wildtype (original) base.
-    Of course, W and SNP could be interchanged here, we don't actually
-    know which is which ;)
-
-    Note our parameterization: we need a bit more room from the ends,
-    so we bring the rightmost SNP a tad left.
-    '''
-
-    graph, wildtype_sequence = linear_structure
-    S = request.param
-    if S < 0:
-        S = len(wildtype_sequence) + S
-    snp_sequence = mutate_position(wildtype_sequence, S + K)
-    HDN_L = Kmer(wildtype_sequence[S:S + K], pos=S)
-    HDN_R = Kmer(wildtype_sequence[S + K + 1:S + 2 * K + 1], pos=S + K + 1)
-
-    graph.consume(wildtype_sequence)
-    graph.consume(snp_sequence)
-
-    # Check for false positive neighbors and mark as expected failure if found
-    w_hdns = hdn_counts(wildtype_sequence, graph)
-    snp_hdns = hdn_counts(snp_sequence, graph)
-    if not (w_hdns == snp_hdns == {3: 2}):
-        print(w_hdns, snp_hdns)
-        print(HDN_L, HDN_R)
-        print(wildtype_sequence[HDN_L.pos + K + 1])
-        print(snp_sequence[HDN_L.pos + K + 1])
-        request.applymarker(pytest.mark.xfail)
-
-    return graph, wildtype_sequence, snp_sequence, HDN_L, HDN_R
-
-
-@pytest.fixture(params=[2, 3, 4, 5, 6, 7, 8])
-def tandem_repeat_structure(request, linear_structure):
-
-    graph, sequence = linear_structure
-
-    tandem_repeats = sequence * request.param
-    graph.consume(tandem_repeats)
-
-    if hdn_counts(tandem_repeats, graph):
-        request.applymarker(pytest.mark.xfail)
-
-    return graph, sequence, tandem_repeats
 
 
 class TestNonBranching:
@@ -537,9 +75,9 @@ class TestNonBranching:
         asm = khmer.LinearAssembler(graph)
 
         for start in range(0, len(contig), 150):
-            path = asm.assemble(contig[start:start + K], direction='L')
+            path = asm.assemble_left(contig[start:start + K])
             print(path, ', ', contig[:start])
-            assert utils._equals_rc(path, contig[:start + K]), start
+            assert utils._equals_rc(path, contig[:start+K]), start
 
     def test_all_right_to_end(self, linear_structure):
         # assemble directed right
@@ -547,7 +85,7 @@ class TestNonBranching:
         asm = khmer.LinearAssembler(graph)
 
         for start in range(0, len(contig), 150):
-            path = asm.assemble(contig[start:start + K], direction='R')
+            path = asm.assemble_right(contig[start:start + K])
             print(path, ', ', contig[:start])
             assert utils._equals_rc(path, contig[start:]), start
 
@@ -574,6 +112,7 @@ class TestLinearAssembler_RightBranching:
         graph, contig, L, HDN, R, tip = right_tip_structure
         asm = khmer.LinearAssembler(graph)
         path = asm.assemble(revcomp(contig[0:K]))
+
 
         assert len(path) == HDN.pos + K
         assert utils._equals_rc(path, contig[:len(path)])
@@ -645,26 +184,28 @@ class TestLinearAssembler_LeftBranching:
         assert utils._equals_rc(path, contig[HDN.pos:])
 
     def test_from_branch_to_ends_with_stopbf(self, left_tip_structure):
-        # block the tip with the stop_bf. should return a full length contig.
+        # block the tip with the stop_filter. should return a full length contig.
         graph, contig, L, HDN, R, tip = left_tip_structure
-        asm = khmer.LinearAssembler(graph)
 
-        stop_bf = khmer.Nodegraph(K, 1e5, 4)
-        stop_bf.count(tip)
+        stop_filter = khmer.Nodegraph(K, 1e5, 4)
+        stop_filter.count(tip)
 
-        path = asm.assemble(HDN, stop_bf)
+        asm = khmer.LinearAssembler(graph, stop_filter=stop_filter)
+
+        path = asm.assemble(HDN)
 
         assert len(path) == len(contig)
         assert utils._equals_rc(path, contig)
 
     def test_from_branch_to_ends_with_stopbf_revcomp(self, left_tip_structure):
-        # block the tip with the stop_bf. should return a full length contig.
+        # block the tip with the stop_filter. should return a full length contig.
         graph, contig, L, HDN, R, tip = left_tip_structure
-        asm = khmer.LinearAssembler(graph)
-        stop_bf = khmer.Nodegraph(K, 1e5, 4)
-        stop_bf.count(tip)
 
-        path = asm.assemble(revcomp(HDN), stop_bf)
+        stop_filter = khmer.Nodegraph(K, 1e5, 4)
+        stop_filter.count(tip)
+        asm = khmer.LinearAssembler(graph, stop_filter=stop_filter)
+
+        path = asm.assemble(revcomp(HDN))
 
         assert len(path) == len(contig)
         assert utils._equals_rc(path, contig)
@@ -673,12 +214,15 @@ class TestLinearAssembler_LeftBranching:
         # assemble up to branch point, and include introduced branch b/c
         # of stop bf
         graph, contig, L, HDN, R, tip = left_tip_structure
-        asm = khmer.LinearAssembler(graph)
 
-        stop_bf = khmer.Nodegraph(K, 1e5, 4)
-        stop_bf.count(L)          # ...and block original path
-        path = asm.assemble(contig[-K:], stop_bf)
+        stop_filter = khmer.Nodegraph(K, 1e5, 4)
+        stop_filter.count(L)          # ...and block original path
+        asm = khmer.LinearAssembler(graph, stop_filter=stop_filter)
+
+
+        path = asm.assemble(contig[-K:])
         assert len(path) == len(contig) - HDN.pos + 1
+
 
         # should be the tip k-kmer, plus the last base of the HDN thru
         # the end of the contig
@@ -830,11 +374,11 @@ class TestLabeledAssembler:
         # assert paths[0][HDN_R.pos:] == paths[1][HDN_R.pos:]
 
     def test_assemble_snp_bubble_stopbf(self, snp_bubble_structure):
-        # assemble one side of bubble, blocked with stop_bf,
+        # assemble one side of bubble, blocked with stop_filter,
         # when labels on both branches
-        # stop_bf should trip a filter failure, negating the label spanning
+        # stop_filter should trip a filter failure, negating the label spanning
         graph, wildtype, mutant, HDN_L, HDN_R = snp_bubble_structure
-        stop_bf = khmer.Nodegraph(K, 1e5, 4)
+        stop_filter = khmer.Nodegraph(K, 1e5, 4)
         lh = khmer._GraphLabels(graph)
         asm = khmer.SimpleLabeledAssembler(lh)
 
@@ -844,9 +388,9 @@ class TestLabeledAssembler:
         lh.label_across_high_degree_nodes(wildtype, hdn, 1)
         lh.label_across_high_degree_nodes(mutant, hdn, 2)
 
-        # do the labeling, but block the mutant with stop_bf
-        stop_bf.count(mutant[HDN_L.pos + 1:HDN_L.pos + K + 1])
-        paths = asm.assemble(wildtype[:K], stop_bf)
+        # do the labeling, but block the mutant with stop_filter
+        stop_filter.count(mutant[HDN_L.pos + 1:HDN_L.pos + K + 1])
+        paths = asm.assemble(wildtype[:K], stop_filter)
 
         assert len(paths) == 1
         assert any(utils._equals_rc(path, wildtype) for path in paths)
