+++ tests/test_cython_assembly.py
@@ -47,7 +47,7 @@ from khmer.khmer_args import estimate_optimal_with_K_and_f as optimal_fp
 from khmer import ReadParser
 from khmer import reverse_complement as revcomp
 from . import khmer_tst_utils as utils
-from khmer._oxli.assembly import LinearAssembler, CompactingAssembler
+from khmer._oxli.assembly import LinearAssembler
 
 import pytest
 import screed
@@ -58,7 +58,7 @@ from .graph_features import K
 def teardown():
     utils.cleanup()
 
-@pytest.mark.parametrize("assembler", [LinearAssembler, CompactingAssembler])
+@pytest.mark.parametrize("assembler", [LinearAssembler])
 class TestNonBranching:
 
     def test_all_start_positions(self, linear_structure, assembler):
@@ -89,28 +89,5 @@ class TestNonBranching:
             path = asm.assemble_right(contig[start:start + K])
             print(path, ', ', contig[:start])
             assert utils._equals_rc(path, contig[start:]), start
-
-
-class TestCompactingAssembler:
-
-    def test_beginning_to_branch_right(self, right_tip_structure):
-        # assemble from beginning of contig, up until branch point
-        graph, contig, L, HDN, R, tip = right_tip_structure
-        asm = CompactingAssembler(graph)
-        path = asm.assemble(contig[0:K])
-
-        assert len(path) == HDN.pos + K
-        assert utils._equals_rc(path, contig[:len(path)])
-
-    def test_end_to_branch_right(self, right_tip_structure):
-        # in the LinearAsembler, this would continue all the way
-        # to the beginning. The CompactingAssembler does an extra
-        # check of the node degree in the reverse direction.
-        graph, contig, L, HDN, R, tip = right_tip_structure
-        asm = CompactingAssembler(graph)
-        path = asm.assemble(contig[-K:])
-
-        assert len(path) == len(contig) - HDN.pos
-        assert utils._equals_rc(path, contig[HDN.pos:])
